Added authentication handling

+++ server/messagehandlers/mongodbhandler.js
@@ -27,7 +27,7 @@ module.exports = (function(db, webSocketServer) {
      */
     async function _handleCreate(params, socket) {
       if (!socket.loggedInUserId) return;
-      params.data.ownerId = socket.loggedInUserId;
+      params.data._ownerId = socket.loggedInUserId;
       const result = await db.create(params.db, params.collection, params.data);
       if (params.collection === 'users') delete result.password;
       return result;
@@ -52,8 +52,8 @@ module.exports = (function(db, webSocketServer) {
      *  data: object
      */
     async function _handleUpdate(params, socket) {
-      //if (!socket.loggedInUserId) return;
-      const result = await db.update(params.db, params.collection, params._id, params.data); // TODO: Filter for ownerId
+      if (!socket.loggedInUserId) return;
+      const result = await db.update(params.db, params.collection, { _id: params._id, _ownerId: socket.loggedInUserId }, params.data);
       if (params.collection === 'users') delete result.password;
       return result;
     }
@@ -65,7 +65,7 @@ module.exports = (function(db, webSocketServer) {
      */
     async function _handleDelete(params, socket) {
       if (!socket.loggedInUserId) return;
-      const result = await db.delete(params.db, params.collection, params._id);
+      const result = await db.delete(params.db, params.collection, { _id: params._id, _ownerId: socket.loggedInUserId });
       if (params.collection === 'users') delete result.password;
       return result;
     }
